pyproject.toml
[project]
name = "ecole"
version = "0.0.1"
description = "eCole walking skeleton"
authors = [{ name = "eCole" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2",
    "PyYAML>=6",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "mypy",
    "ruff",
    "black",
    "types-PyYAML",
]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.11"
warn_unused_ignores = true
disallow_untyped_defs = true
no_implicit_optional = true
